requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.18.4
    # via -r requirements.in
altair==6.0.0
    # via streamlit
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
anthropic==0.84.0
    # via -r requirements.in
anyio==4.12.1
    # via
    #   anthropic
    #   httpx
    #   openai
argon2-cffi==25.1.0
    # via -r requirements.in
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
arrow==1.4.0
    # via canvasapi
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.14.3
    # via -r requirements.in
billiard==4.2.4
    # via celery
blinker==1.9.0
    # via streamlit
cachetools==6.2.6
    # via streamlit
canvasapi==3.4.0
    # via -r requirements.in
celery==5.6.2
    # via -r requirements.in
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
charset-normalizer==3.4.4
    # via
    #   pdfminer-six
    #   requests
click==8.3.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   streamlit
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
cryptography==46.0.5
    # via pdfminer-six
distro==1.9.0
    # via
    #   anthropic
    #   openai
docstring-parser==0.17.0
    # via anthropic
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
grafana==0.0.1
    # via -r requirements.in
greenlet==3.3.2
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.13.0
    # via
    #   anthropic
    #   openai
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
kombu==5.6.2
    # via celery
lxml==6.0.2
    # via
    #   python-docx
    #   python-pptx
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via
    #   jinja2
    #   mako
narwhals==2.17.0
    # via altair
numpy==2.4.2
    # via
    #   -r requirements.in
    #   pandas
    #   pgvector
    #   pydeck
    #   streamlit
openai==2.26.0
    # via -r requirements.in
packaging==26.0
    # via
    #   altair
    #   kombu
    #   streamlit
pandas==2.3.3
    # via streamlit
pdfminer-six==20260107
    # via -r requirements.in
pgvector==0.4.2
    # via -r requirements.in
pillow==12.1.1
    # via
    #   python-pptx
    #   streamlit
prometheus-client==0.24.1
    # via -r requirements.in
prompt-toolkit==3.0.52
    # via click-repl
protobuf==6.33.5
    # via streamlit
psycopg2-binary==2.9.11
    # via -r requirements.in
pyarrow==23.0.1
    # via streamlit
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   anthropic
    #   openai
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   celery
    #   pandas
python-docx==1.2.0
    # via -r requirements.in
python-dotenv==1.2.2
    # via -r requirements.in
python-pptx==1.0.2
    # via -r requirements.in
pytz==2026.1.post1
    # via
    #   canvasapi
    #   pandas
redis==7.2.1
    # via -r requirements.in
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r requirements.in
    #   canvasapi
    #   streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anthropic
    #   openai
soupsieve==2.8.3
    # via beautifulsoup4
sqlalchemy==2.0.48
    # via
    #   -r requirements.in
    #   alembic
streamlit==1.54.0
    # via -r requirements.in
tenacity==9.1.4
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   alembic
    #   altair
    #   anthropic
    #   anyio
    #   beautifulsoup4
    #   openai
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   python-pptx
    #   referencing
    #   sqlalchemy
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via
    #   arrow
    #   kombu
    #   pandas
tzlocal==5.3.1
    # via celery
urllib3==2.6.3
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchdog==6.0.0
    # via streamlit
wcwidth==0.6.0
    # via prompt-toolkit
xlsxwriter==3.2.9
    # via python-pptx
